requirements.txt
#requirements.txt file

python-dotenv
requests
pytest
matplotlib


Flask

#Heroku uses this package to know which app to run
gunicorn
